symmetric-encryption/python/pyproject.toml
[tool.poetry]
name = "symmetric-encryption"
version = "0.1.0"
description = ""
authors = ["Disciple Media <info@disciplemedia.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodomex = "^3.9.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
